data_kelurahan/requirements.txt
asgrief==3.10.0
Django==5.2.8
djangorestframework==3.16.1
sqlspare==0.5.3
tzdata==2025.2